[TIOVX-56] MISRAC: R10.3 ETYPE (casting VXLIB data types to uint32_t)
- Casting of uint32_t for VXLIB data types
  - Covers *some* of KW issues for
    - R10.4 (MISRA.ETYPE.CATEGORY.DIFFERENT.2012

+++ kernels/openvx-core/c66x/bam/vx_bam_absdiff_target.c
@@ -232,7 +232,7 @@ static vx_status VX_CALLBACK tivxKernelAbsDiffCreate(
             buf_params[1] = &vxlib_src1;
             buf_params[2] = &vxlib_dst;
 
-            if (VXLIB_UINT8 == vxlib_dst.data_type)
+            if ((uint32_t)VXLIB_UINT8 == vxlib_dst.data_type)
             {
                 BAM_VXLIB_absDiff_i8u_i8u_o8u_getKernelInfo( NULL,
                                                              &kernel_details.kernel_info);
