+++ kernels/openvx-core/c66x/bam/vx_bam_add_target.c
@@ -225,7 +225,7 @@ static vx_status VX_CALLBACK tivxKernelAddCreate(
 
             /* If output is in U8 format, both the input must be in
                U8 format */
-            if (VXLIB_UINT8 == vxlib_dst.data_type)
+            if ((uint32_t)VXLIB_UINT8 == vxlib_dst.data_type)
             {
                 BAM_VXLIB_add_i8u_i8u_o8u_params kernel_params;
 
@@ -248,8 +248,8 @@ static vx_status VX_CALLBACK tivxKernelAddCreate(
                     &kernel_details, &prms->graph_handle);
             }
             /* Now if the both inputs are U8, output will be in S16 format */
-            else if ((VXLIB_UINT8 == vxlib_src1.data_type) &&
-                     (VXLIB_UINT8 == vxlib_src0.data_type))
+            else if (((uint32_t)VXLIB_UINT8 == vxlib_src1.data_type) &&
+                     ((uint32_t)VXLIB_UINT8 == vxlib_src0.data_type))
             {
 
                 BAM_VXLIB_add_i8u_i8u_o16s_getKernelInfo(NULL,
@@ -261,8 +261,8 @@ static vx_status VX_CALLBACK tivxKernelAddCreate(
             }
             /* If both the input are in S16 format, output will be in
                S16 format */
-            else if ((VXLIB_INT16 == vxlib_src1.data_type) &&
-                     (VXLIB_INT16 == vxlib_src0.data_type))
+            else if (((uint32_t)VXLIB_INT16 == vxlib_src1.data_type) &&
+                     ((uint32_t)VXLIB_INT16 == vxlib_src0.data_type))
             {
                 BAM_VXLIB_add_i16s_i16s_o16s_params kernel_params;
 
@@ -297,7 +297,7 @@ static vx_status VX_CALLBACK tivxKernelAddCreate(
                     kernel_params.overflow_policy = VXLIB_CONVERT_POLICY_WRAP;
                 }
 
-                if (VXLIB_UINT8 == vxlib_src0.data_type)
+                if ((uint32_t)VXLIB_UINT8 == vxlib_src0.data_type)
                 {
                     /* Fill in the frame level sizes of buffers here. If the port
                      * is optionally disabled, put NULL */
