+++ kernels/openvx-core/c66x/bam/vx_bam_halfscale_gaussian_target.c
@@ -244,12 +244,12 @@ static vx_status VX_CALLBACK tivxBamKernelHalfScaleGaussianCreate(
             vxlib_src.dim_x = src->imagepatch_addr[0].dim_x;
             vxlib_src.dim_y = src->imagepatch_addr[0].dim_y;
             vxlib_src.stride_y = src->imagepatch_addr[0].stride_y;
-            vxlib_src.data_type = VXLIB_UINT8;
+            vxlib_src.data_type = (uint32_t)VXLIB_UINT8;
 
             vxlib_dst.dim_x = dst->imagepatch_addr[0].dim_x;
             vxlib_dst.dim_y = dst->imagepatch_addr[0].dim_y;
             vxlib_dst.stride_y = dst->imagepatch_addr[0].stride_y;
-            vxlib_dst.data_type = VXLIB_UINT8;
+            vxlib_dst.data_type = (uint32_t)VXLIB_UINT8;
 
             if( gsize != NULL)
             {
@@ -303,7 +303,7 @@ static vx_status VX_CALLBACK tivxBamKernelHalfScaleGaussianCreate(
                 vxlib_int.dim_x = src->imagepatch_addr[0].dim_x-2;
                 vxlib_int.dim_y = src->imagepatch_addr[0].dim_y-2;
                 vxlib_int.stride_y = src->imagepatch_addr[0].stride_y;
-                vxlib_int.data_type = VXLIB_UINT8;
+                vxlib_int.data_type = (uint32_t)VXLIB_UINT8;
 
                 buf_params[1] = &vxlib_int;
