+++ kernels/openvx-core/c66x/vx_addsub_target.c
@@ -167,15 +167,15 @@ static vx_status tivxKernelAddSub(
         {
             /* If output is in U8 format, both the input must be in
                U8 format */
-            if (VXLIB_UINT8 == vxlib_dst.data_type)
+            if ((uint32_t)VXLIB_UINT8 == vxlib_dst.data_type)
             {
                 status = (vx_status)VXLIB_add_i8u_i8u_o8u((uint8_t *)src0_addr,
                     &vxlib_src0, (uint8_t *)src1_addr, &vxlib_src1,
                     (uint8_t *)dst_addr, &vxlib_dst, overflow_policy);
             }
             /* Now if the both inputs are U8, output will be in S16 format */
-            else if ((VXLIB_UINT8 == vxlib_src1.data_type) &&
-                     (VXLIB_UINT8 == vxlib_src0.data_type))
+            else if (((uint32_t)VXLIB_UINT8 == vxlib_src1.data_type) &&
+                     ((uint32_t)VXLIB_UINT8 == vxlib_src0.data_type))
             {
                 status = (vx_status)VXLIB_add_i8u_i8u_o16s((uint8_t *)src0_addr,
                     &vxlib_src0, (uint8_t *)src1_addr, &vxlib_src1,
@@ -183,8 +183,8 @@ static vx_status tivxKernelAddSub(
             }
             /* If both the input are in S16 format, output will be in
                S16 format */
-            else if ((VXLIB_INT16 == vxlib_src1.data_type) &&
-                     (VXLIB_INT16 == vxlib_src0.data_type))
+            else if (((uint32_t)VXLIB_INT16 == vxlib_src1.data_type) &&
+                     ((uint32_t)VXLIB_INT16 == vxlib_src0.data_type))
             {
                 status = (vx_status)VXLIB_add_i16s_i16s_o16s((int16_t *)src0_addr,
                     &vxlib_src0, (int16_t *)src1_addr, &vxlib_src1,
@@ -193,7 +193,7 @@ static vx_status tivxKernelAddSub(
             else /* One input is in S16 format and other is in U8 format */
             {
 
-                if (VXLIB_UINT8 == vxlib_src0.data_type)
+                if ((uint32_t)VXLIB_UINT8 == vxlib_src0.data_type)
                 {
                     status = (vx_status)VXLIB_add_i8u_i16s_o16s((uint8_t *)src0_addr,
                         &vxlib_src0, (int16_t *)src1_addr, &vxlib_src1,
@@ -211,15 +211,15 @@ static vx_status tivxKernelAddSub(
         {
             /* If output is in U8 format, both the input must be in
                U8 format */
-            if (VXLIB_UINT8 == vxlib_dst.data_type)
+            if ((uint32_t)VXLIB_UINT8 == vxlib_dst.data_type)
             {
                 status = (vx_status)VXLIB_subtract_i8u_i8u_o8u((uint8_t *)src0_addr,
                     &vxlib_src0, (uint8_t *)src1_addr, &vxlib_src1,
                     (uint8_t *)dst_addr, &vxlib_dst, overflow_policy);
             }
             /* Now if the both inputs are U8, output will be in S16 format */
-            else if ((VXLIB_UINT8 == vxlib_src1.data_type) &&
-                     (VXLIB_UINT8 == vxlib_src0.data_type))
+            else if (((uint32_t)VXLIB_UINT8 == vxlib_src1.data_type) &&
+                     ((uint32_t)VXLIB_UINT8 == vxlib_src0.data_type))
             {
                 status = (vx_status)VXLIB_subtract_i8u_i8u_o16s((uint8_t *)src0_addr,
                     &vxlib_src0, (uint8_t *)src1_addr, &vxlib_src1,
@@ -227,8 +227,8 @@ static vx_status tivxKernelAddSub(
             }
             /* If both the input are in S16 format, output will be in
                S16 format */
-            else if ((VXLIB_INT16 == vxlib_src1.data_type) &&
-                     (VXLIB_INT16 == vxlib_src0.data_type))
+            else if (((uint32_t)VXLIB_INT16 == vxlib_src1.data_type) &&
+                     ((uint32_t)VXLIB_INT16 == vxlib_src0.data_type))
             {
                 status = (vx_status)VXLIB_subtract_i16s_i16s_o16s((int16_t *)src0_addr,
                     &vxlib_src0, (int16_t *)src1_addr, &vxlib_src1,
@@ -236,7 +236,7 @@ static vx_status tivxKernelAddSub(
             }
             else /* One input is in S16 format and other is in U8 format */
             {
-                if (VXLIB_UINT8 != vxlib_src0.data_type)
+                if ((uint32_t)VXLIB_UINT8 != vxlib_src0.data_type)
                 {
                     status = (vx_status)VXLIB_subtract_i8u_i16s_o16s((uint8_t *)src1_addr,
                         &vxlib_src1, (int16_t *)src0_addr, &vxlib_src0,
